+++ examples/stdout.py
@@ -6,16 +6,14 @@ from __future__ import unicode_literals
 import sys
 import time
 
-from eliot import Message, Logger, to_file
+from eliot import Message, to_file
 to_file(sys.stdout)
-
-_logger = Logger()
 
 
 def main():
-    Message.new(value="hello", another=1).write(_logger)
+    Message.new(value="hello", another=1).write()
     time.sleep(0.2)
-    Message.new(value="goodbye", another=2).write(_logger)
+    Message.new(value="goodbye", another=2).write()
 
 
 if __name__ == '__main__':