+++ examples/stdout.py
@@ -11,9 +11,9 @@ to_file(sys.stdout)
 
 
 def main():
-    Message.new(value="hello", another=1).write()
+    Message.log(value="hello", another=1)
     time.sleep(0.2)
-    Message.new(value="goodbye", another=2).write()
+    Message.log(value="goodbye", another=2)
 
 
 if __name__ == '__main__':